Moves badge_achievement-related code from /admin/bagdes to /admin/badge_achievements   - Adds an Admin::BadgeAchievements::Controller   - Moves #award and #award_badges to Admin::BadgeAchievements::Controller   - Removes #award and #award_badges from Admin::Badges::Controller   - Moves award.html.erb from /admin/badges to /admin/badge_achievements   - Removes badge_achievement routes from /admin/badges   - Adds a redirect for /admin/badges/badge_achievements to routes.rb   - Cleans up and refactors code in controllers, views, and routes
