+++ config/routes.rb
@@ -117,8 +117,9 @@ Rails.application.routes.draw do
     resources :webhook_endpoints, only: :index
     resource :config
     resources :badges, only: %i[index edit update new create]
-    get "/badge_achievements/award_badges", to: "badges#award"
-    post "/badge_achievements/award_badges", to: "badges#award_badges"
+    get "/badge_achievements/award_badges", to: redirect("/admin/badge_achievements")
+    get "/badge_achievements", to: "badge_achievements#award"
+    post "/badge_achievements/award_badges", to: "badge_achievements#award_badges"
     resources :secrets, only: %i[index]
     put "secrets", to: "secrets#update"
   end
